Flatten CLI
